requirements.txt
pdf2image==1.16.3
pdfminer.six==20221105
pyzbar==0.1.9